requirements.txt
streamlit
torch
torchvision
Pillow
gdown>=4.7.1